+++ ProgrammationBBB/BBBEurobot/include/BBBEurobot/PCAServoDriver.h
@@ -6,14 +6,14 @@
 #ifndef PCASERVO_DRIVER
 	#define PCASERVO_DRIVER
 	#include "BBBEurobot/I2C-Wrapper.h"
-	
+
 	/// Led driver structure.
 	typedef struct {
 		I2CAdapter *adapter;		///< I2C port file descriptor.
-		int address;	///< Servo driver address.
+		guint8 address;	///< Servo driver address.
 	}ServoDriver;
-	
-	
+
+
 	/// \brief Initialize the servo driver.
     ///
     /// \details This function tests the communication with the servo driver, and, if successful, inits the structure.
@@ -23,13 +23,13 @@
     ///                    see I2C-Wrapper.h).
     /// \param[in] address I2C address of the ServoDriver.
     /// \returns   TRUE on success, FALSE otherwise.
-	gboolean servoDriver_init(ServoDriver *driver, I2CAdapter *adapter, int address);	
-	
+	gboolean servoDriver_init(ServoDriver *driver, I2CAdapter *adapter, guint8 address);
+
 	/// \brief Set position of a servo.
     ///
     /// \param[in,out] driver The PCA9635 structure (as a pointer as the ledState variable might be modified).
     /// \param[in] servo The number of the servo to change (from 0 to 15).
     /// \param[in] position Signal value, in microseconds (clamped between 500 and 2500). Note that the resolution
     ///                   of the driver is of 4us only.
-	void servoDriver_setPosition(ServoDriver driver, int servo, int position);	
+	void servoDriver_setPosition(ServoDriver driver, int servo, int position);
 #endif
